chore(clippy): zero pre-existing warnings — Lite-W6 PR-0c
48 errors → 0 across lib + tests. cargo clippy --all-targets -- -D warnings
is now a real lock instead of a deferred gate.

- Auto-fix (~28): useless_conversion, needless_question_mark,
  unnecessary_map_or, unnecessary_cast, unnecessary_parens, for_kv_map,
  needless_borrow, unused imports, unused mut.
- Manual: explicit type on std::mem::transmute (db.rs:71);
  let...else → ? (indexer.rs:412).
- Dead code deleted: insert_session (sync test helper, 0 callers;
  the suspected await-missing was this dead wrapper, not real code) +
  _suppress_unused; LlmError::MissingApiKey (0 constructs, 0 matches).
- #[allow]: clippy::too_many_arguments × 12 (tauri::command IPC
  signatures + Db CRUD — refactor out of scope for PR-0c);
  dead_code on EMBEDDING_DIM + LlmProvider::name (intentional API
  surface; 4 impls of name() exist).

Verified 5/5 green: pnpm typecheck/lint/test + cargo test (228) +
cargo clippy --all-targets -- -D warnings.

Docs: master-prompt §5.1/§5.2 + 07-checklist PR0 clippy ⚠→☑.

+++ src-tauri/src/oculpm/migrate_from_sqlite.rs
@@ -1108,6 +1108,7 @@ mod tests {
     }
 
     /// Insert one entry at `base_ts + offset_secs`. Returns the new id.
+    #[allow(clippy::too_many_arguments)]
     async fn seed(
         db: &Db,
         project_id: u32,
@@ -1138,7 +1139,7 @@ mod tests {
         db.conn()
             .call({
                 let id = entry.id as i64;
-                let ts = (base_ts + offset_secs) as i64;
+                let ts = base_ts + offset_secs;
                 move |c| {
                     c.execute(
                         "UPDATE changelog_entries SET created_at = ?1 WHERE id = ?2",
@@ -1636,7 +1637,7 @@ mod tests {
         db.conn()
             .call({
                 let id = entry.id as i64;
-                let ts = (base + 3600) as i64;
+                let ts = base + 3600;
                 move |c| {
                     c.execute(
                         "UPDATE changelog_entries SET created_at = ?1 WHERE id = ?2",
